// react-module/src/openapi/models/erp-order.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface ErpOrder
 */
export interface ErpOrder {
    /**
     * 
     * @type {Gid}
     * @memberof ErpOrder
     */
    gid?: any;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    creationDate?: any;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    executionDate?: any;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    type?: ErpOrderTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    subType?: ErpOrderSubTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    orderNumber?: any;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    foreignDocument?: any;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    description?: any;
    /**
     * 
     * @type {string}
     * @memberof ErpOrder
     */
    state?: ErpOrderStateEnum;
    /**
     * 
     * @type {ContractorData}
     * @memberof ErpOrder
     */
    contractorData?: any;
    /**
     * 
     * @type {WarehouseDefinition}
     * @memberof ErpOrder
     */
    warehouse?: any;
    /**
     * 
     * @type {WarehouseDefinition}
     * @memberof ErpOrder
     */
    targetWarehouse?: any;
    /**
     * 
     * @type {UserDefinition}
     * @memberof ErpOrder
     */
    owner?: any;
    /**
     * 
     * @type {number}
     * @memberof ErpOrder
     */
    expoNorm?: any;
}

/**
    * @export
    * @enum {string}
    */
export enum ErpOrderTypeEnum {
    SALE = 'SALE',
    PURCHASE = 'PURCHASE',
    PURCHASEOFFER = 'PURCHASE_OFFER',
    SALEOFFER = 'SALE_OFFER',
    UNKNOWN = 'UNKNOWN'
}
/**
    * @export
    * @enum {string}
    */
export enum ErpOrderSubTypeEnum {
    INTERNAL = 'INTERNAL',
    EXTERNAL = 'EXTERNAL'
}
/**
    * @export
    * @enum {string}
    */
export enum ErpOrderStateEnum {
    QUERY = 'QUERY',
    OFFER = 'OFFER',
    UNCONFIRMED = 'UNCONFIRMED',
    CONFIRMED = 'CONFIRMED',
    ACCEPTED = 'ACCEPTED',
    INEXECUTION = 'IN_EXECUTION',
    REJECTED = 'REJECTED',
    COMPLETED = 'COMPLETED',
    CANCELEDCONFIRMED = 'CANCELED_CONFIRMED',
    UNKNOWN = 'UNKNOWN'
}